requirements.txt
Flask
pytesseract
pillow
sqlalchemy
